metamorph: Expose that the new label goes at the head during fold.
For any valid Row, the Forall instance is such that the new label
being added during the fold step of a metamorphism belongs at the head
of the LT list if injected (i.e. the list so far does not contain the
new label and the new label sorts before the current head of the
list), since the fold just rebuilds the original Row in its
already-sorted and unique state. Exposing this via the FoldStep
constraint allows consumers of row-types to build up their structures
without accessing library internals.

In principle this should allow us to remove e.g. the unsafeInjectFront
calls in the implementation of rmapc, but that would require proving
that Map preserves FoldStep first.